info-pretty: Avoid repeated face remaps
Calling `face-remap-add-relative` unconditionally from `Info-selection-hook`
adds an entry to `face-remapping-alist` every time a section is opened.  This
is especially problematic if, for example, the 'info-pretty-body' face uses
relative font heights, which end up being multiplied together.

Instead, save the cookie returned by `face-remap-add-relative` as a
buffer-local variable and check for it before adding the face remap again.

Also, adding a face remapping of the form `(:inherit face)` is exactly
equivalent to just remapping to 'face', so do that instead.